feat: updated colors docs page to display updated tokens
