+++ package.json
@@ -2,7 +2,7 @@
   "name": "@casavo/habitat",
   "description": "Casavo Design System Library",
   "private": false,
-  "version": "1.0.0-beta.12",
+  "version": "1.0.0-beta.13",
   "type": "module",
   "license": "Apache-2.0",
   "homepage": "https://github.com/casavo/habitat",
@@ -34,31 +34,31 @@
     "test:unit": "vitest run"
   },
   "dependencies": {
-    "@vanilla-extract/css": "1.14.1",
+    "@vanilla-extract/css": "1.14.2",
     "@vanilla-extract/recipes": "0.5.2",
     "react": "18.2.0",
     "react-aria-components": "1.0.0-beta.1",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "8.0.4",
-    "@storybook/addon-interactions": "8.0.4",
-    "@storybook/addon-links": "8.0.4",
-    "@storybook/blocks": "8.0.4",
-    "@storybook/react": "8.0.4",
-    "@storybook/react-vite": "8.0.4",
+    "@storybook/addon-essentials": "8.0.6",
+    "@storybook/addon-interactions": "8.0.6",
+    "@storybook/addon-links": "8.0.6",
+    "@storybook/blocks": "8.0.6",
+    "@storybook/react": "8.0.6",
+    "@storybook/react-vite": "8.0.6",
     "@storybook/testing-library": "0.2.2",
     "@testing-library/jest-dom": "6.4.2",
     "@testing-library/react": "14.2.2",
     "@types/jest": "29.5.12",
     "@types/mocha": "10.0.6",
-    "@types/react": "18.2.72",
-    "@types/react-dom": "18.2.22",
-    "@typescript-eslint/eslint-plugin": "7.4.0",
-    "@typescript-eslint/parser": "7.4.0",
-    "@vanilla-extract/vite-plugin": "4.0.6",
+    "@types/react": "18.2.74",
+    "@types/react-dom": "18.2.24",
+    "@typescript-eslint/eslint-plugin": "7.5.0",
+    "@typescript-eslint/parser": "7.5.0",
+    "@vanilla-extract/vite-plugin": "4.0.7",
     "@vitejs/plugin-react": "4.2.1",
-    "chromatic": "11.2.0",
+    "chromatic": "11.3.0",
     "eslint": "8.57.0",
     "eslint-plugin-react": "7.34.1",
     "eslint-plugin-react-hooks": "4.6.0",
@@ -71,11 +71,11 @@
     "prop-types": "15.8.1",
     "react-aria": "3.32.1",
     "react-aria-components": "1.0.0-beta.1",
-    "storybook": "8.0.4",
+    "storybook": "8.0.6",
     "style-dictionary": "3.9.2",
-    "typescript": "5.4.3",
-    "vite": "5.2.6",
-    "vite-plugin-dts": "3.7.3",
+    "typescript": "5.4.4",
+    "vite": "5.2.8",
+    "vite-plugin-dts": "3.8.1",
     "vite-tsconfig-paths": "4.3.2",
     "vitest": "1.4.0"
   }
