+++ src/index.ts
@@ -4,3 +4,7 @@ export { Button } from "./components/Button";
 export { Badge } from "./components/Badge";
 export { Switch } from "./components/Switch";
 export { Checkbox } from "./components/Checkbox";
+export { colors } from "./utils/colors";
+export { corners } from "./utils/corners";
+export { space } from "./utils/spacing";
+export { mq } from "./utils/mediaqueries";
