+++ src/stories/Colors.Semantic.mdx
@@ -28,5 +28,10 @@ import { vars } from "./../utils/theme.css";
     <Body>Foreground is essentially applied to any text, icons, and any UI elements that sit on top, or on the front, of a background.</Body>
     <ColorSection values="foreground" />
 
+    <hr style={{ marginTop: vars.space[1000] }} />
+    <H2>Buttons (interactive)</H2>
+    <Body>Most commonly, “interactive” are your buttons. These are further broken down into hierarchies—primary, secondary, and tertiary. Each has their own states — <Inline strong>default</Inline>, <Inline strong>hover</Inline>, <Inline strong>active</Inline>, <Inline strong>disabled</Inline>, and <Inline strong>inverted</Inline> (for dark mode).</Body>
+    <ColorSection values="buttons.light" />
+
   </StoryLayout>
 </div>
