+++ src/stories/Foundations.Corners.mdx
@@ -8,8 +8,6 @@ import {
   H5,
   H6,
   Body,
-  Description,
-  Caption,
 } from "../components/Typography";
 import { Table } from "./components/Corners";
