+++ src/stories/Foundations.Spacings.mdx
@@ -8,8 +8,6 @@ import {
   H5,
   H6,
   Body,
-  Description,
-  Caption,
 } from "../components/Typography";
 import { Table } from "./components/Spacings";
